add memorycache.cpp; move associated generic functions/structs to utils
