+++ src/binding.cpp
@@ -1,7 +1,4 @@
-
 #include "binding.hpp"
-#include "node_util.hpp"
-#include "coalesce.hpp"
 
 #include <algorithm>
 #include <cassert>
@@ -20,80 +17,11 @@ namespace carmen {
 
 using namespace v8;
 
-Nan::Persistent<FunctionTemplate> MemoryCache::constructor;
+
 Nan::Persistent<FunctionTemplate> RocksDBCache::constructor;
 Nan::Persistent<FunctionTemplate> NormalizationCache::constructor;
 
-rocksdb::Status OpenDB(const rocksdb::Options& options, const std::string& name, std::unique_ptr<rocksdb::DB>& dbptr) {
-    rocksdb::DB* db;
-    rocksdb::Status status = rocksdb::DB::Open(options, name, &db);
-    dbptr = std::unique_ptr<rocksdb::DB>(db);
-    return status;
-}
 
-rocksdb::Status OpenForReadOnlyDB(const rocksdb::Options& options, const std::string& name, std::unique_ptr<rocksdb::DB>& dbptr) {
-    rocksdb::DB* db;
-    rocksdb::Status status = rocksdb::DB::OpenForReadOnly(options, name, &db);
-    dbptr = std::unique_ptr<rocksdb::DB>(db);
-    return status;
-}
-
-// in general, the key format including language field is <key_text><separator character><8-byte langfield>
-// but as a size optimization for language-less indexes, we omit the langfield
-// if it would otherwise have been ALL_LANGUAGES (all 1's), so if the first occurence
-// of the separator character is also the last character in the string, we just retur ALL_LANGUAGES
-// otherwise we extract it from the string
-//
-// we centralize both the adding of the field and extracting of the field here to keep from having
-// to handle that optimization everywhere
-inline langfield_type extract_langfield(std::string const& s) {
-    size_t length = s.length();
-    size_t langfield_start = s.find(LANGFIELD_SEPARATOR) + 1;
-    size_t distance_from_end = length - langfield_start;
-
-    if (distance_from_end == 0) {
-        return ALL_LANGUAGES;
-    } else {
-        langfield_type result(0);
-        memcpy(&result, s.data() + langfield_start, distance_from_end);
-        return result;
-    }
-}
-
-inline void add_langfield(std::string& s, langfield_type langfield) {
-    if (langfield != ALL_LANGUAGES) {
-        char* lf_as_char = reinterpret_cast<char*>(&langfield);
-
-        // we only want to copy over as many bytes as we're using
-        // so find the last byte that's not zero, and copy until there
-        // NOTE: this assumes little-endianness, where the bytes
-        // in use will be first rather than last
-        size_t highest(0);
-        for (size_t i = 0; i < sizeof(langfield_type); i++) {
-            if (lf_as_char[i] != 0) highest = i;
-        }
-        size_t field_length = highest + 1;
-
-        s.reserve(sizeof(LANGFIELD_SEPARATOR) + field_length);
-        s.push_back(LANGFIELD_SEPARATOR);
-        s.append(lf_as_char, field_length);
-    } else {
-        s.push_back(LANGFIELD_SEPARATOR);
-    }
-}
-
-intarray __get(MemoryCache const* c, std::string phrase, langfield_type langfield) {
-    arraycache const& cache = c->cache_;
-    intarray array;
-
-    add_langfield(phrase, langfield);
-    arraycache::const_iterator aitr = cache.find(phrase);
-    if (aitr != cache.end()) {
-        array = aitr->second;
-    }
-    std::sort(array.begin(), array.end(), std::greater<uint64_t>());
-    return array;
-}
 
 inline void decodeMessage(std::string const& message, intarray& array) {
     protozero::pbf_reader item(message);
@@ -149,38 +77,6 @@ struct sortableGrid {
     value_type unadjusted_lastval;
     bool matches_language;
 };
-
-intarray __getmatching(MemoryCache const* c, std::string phrase, bool match_prefixes, langfield_type langfield) {
-    intarray array;
-
-    if (!match_prefixes) phrase.push_back(LANGFIELD_SEPARATOR);
-    size_t phrase_length = phrase.length();
-    const char* phrase_data = phrase.data();
-
-    // Load values from memory cache
-
-    for (auto const& item : c->cache_) {
-        const char* item_data = item.first.data();
-        size_t item_length = item.first.length();
-
-        if (item_length < phrase_length) continue;
-
-        if (memcmp(phrase_data, item_data, phrase_length) == 0) {
-            langfield_type message_langfield = extract_langfield(item.first);
-
-            if (message_langfield & langfield) {
-                array.reserve(array.size() + item.second.size());
-                for (auto const& grid : item.second) {
-                    array.emplace_back(grid | LANGUAGE_MATCH_BOOST);
-                }
-            } else {
-                array.insert(array.end(), item.second.begin(), item.second.end());
-            }
-        }
-    }
-    std::sort(array.begin(), array.end(), std::greater<uint64_t>());
-    return array;
-}
 
 intarray __getmatching(RocksDBCache const* c, std::string phrase, bool match_prefixes, langfield_type langfield) {
     intarray array;
@@ -264,25 +160,7 @@ intarray __getmatching(RocksDBCache const* c, std::string phrase, bool match_pre
     return array;
 }
 
-void MemoryCache::Initialize(Handle<Object> target) {
-    Nan::HandleScope scope;
-    Local<FunctionTemplate> t = Nan::New<FunctionTemplate>(MemoryCache::New);
-    t->InstanceTemplate()->SetInternalFieldCount(1);
-    t->SetClassName(Nan::New("MemoryCache").ToLocalChecked());
-    Nan::SetPrototypeMethod(t, "pack", MemoryCache::pack);
-    Nan::SetPrototypeMethod(t, "list", MemoryCache::list);
-    Nan::SetPrototypeMethod(t, "_set", _set);
-    Nan::SetPrototypeMethod(t, "_get", _get);
-    Nan::SetPrototypeMethod(t, "_getMatching", _getmatching);
-    target->Set(Nan::New("MemoryCache").ToLocalChecked(), t->GetFunction());
-    constructor.Reset(t);
-}
 
-MemoryCache::MemoryCache()
-    : ObjectWrap(),
-      cache_() {}
-
-MemoryCache::~MemoryCache() {}
 
 void RocksDBCache::Initialize(Handle<Object> target) {
     Nan::HandleScope scope;
@@ -303,135 +181,6 @@ RocksDBCache::RocksDBCache()
       db() {}
 
 RocksDBCache::~RocksDBCache() {}
-
-inline void packVec(intarray const& varr, std::unique_ptr<rocksdb::DB> const& db, std::string const& key) {
-    std::string message;
-
-    protozero::pbf_writer item_writer(message);
-
-    {
-        // Using new (in protozero 1.3.0) packed writing API
-        // https://github.com/mapbox/protozero/commit/4e7e32ac5350ea6d3dcf78ff5e74faeee513a6e1
-        protozero::packed_field_uint64 field{item_writer, 1};
-        uint64_t lastval = 0;
-        for (auto const& vitem : varr) {
-            if (lastval == 0) {
-                field.add_element(static_cast<uint64_t>(vitem));
-            } else {
-                field.add_element(static_cast<uint64_t>(lastval - vitem));
-            }
-            lastval = vitem;
-        }
-    }
-
-    db->Put(rocksdb::WriteOptions(), key, message);
-}
-
-NAN_METHOD(MemoryCache::pack) {
-    if (info.Length() < 1) {
-        return Nan::ThrowTypeError("expected one info: 'filename'");
-    }
-    if (!info[0]->IsString()) {
-        return Nan::ThrowTypeError("first argument must be a String");
-    }
-    try {
-        Nan::Utf8String utf8_filename(info[0]);
-        if (utf8_filename.length() < 1) {
-            return Nan::ThrowTypeError("first arg must be a String");
-        }
-        std::string filename(*utf8_filename);
-
-        MemoryCache* c = node::ObjectWrap::Unwrap<MemoryCache>(info.This());
-
-        std::unique_ptr<rocksdb::DB> db;
-        rocksdb::Options options;
-        options.create_if_missing = true;
-        rocksdb::Status status = OpenDB(options, filename, db);
-
-        if (!status.ok()) {
-            return Nan::ThrowTypeError("unable to open rocksdb file for packing");
-        }
-
-        std::map<key_type, std::deque<value_type>> memoized_prefixes;
-
-        for (auto const& item : c->cache_) {
-            std::size_t array_size = item.second.size();
-            if (array_size > 0) {
-                // make copy of intarray so we can sort without
-                // modifying the original array
-                intarray varr = item.second;
-
-                // delta-encode values, sorted in descending order.
-                std::sort(varr.begin(), varr.end(), std::greater<uint64_t>());
-
-                packVec(varr, db, item.first);
-
-                std::string prefix_t1 = "";
-                std::string prefix_t2 = "";
-
-                // add this to the memoized prefix array too, maybe
-                auto phrase_length = item.first.find(LANGFIELD_SEPARATOR);
-                // use the full string for things shorter than the limit
-                // or the prefix otherwise
-                if (phrase_length < MEMO_PREFIX_LENGTH_T1) {
-                    prefix_t1 = "=1" + item.first;
-                } else {
-                    // get the prefix, then append the langfield back onto it again
-                    langfield_type langfield = extract_langfield(item.first);
-
-                    prefix_t1 = "=1" + item.first.substr(0, MEMO_PREFIX_LENGTH_T1);
-                    add_langfield(prefix_t1, langfield);
-
-                    if (phrase_length < MEMO_PREFIX_LENGTH_T2) {
-                        prefix_t2 = "=2" + item.first;
-                    } else {
-                        prefix_t2 = "=2" + item.first.substr(0, MEMO_PREFIX_LENGTH_T2);
-                        add_langfield(prefix_t2, langfield);
-                    }
-                }
-
-                if (prefix_t1 != "") {
-                    std::map<key_type, std::deque<value_type>>::const_iterator mitr = memoized_prefixes.find(prefix_t1);
-                    if (mitr == memoized_prefixes.end()) {
-                        memoized_prefixes.emplace(prefix_t1, std::deque<value_type>());
-                    }
-                    std::deque<value_type>& buf = memoized_prefixes[prefix_t1];
-
-                    buf.insert(buf.end(), varr.begin(), varr.end());
-                }
-                if (prefix_t2 != "") {
-                    std::map<key_type, std::deque<value_type>>::const_iterator mitr = memoized_prefixes.find(prefix_t2);
-                    if (mitr == memoized_prefixes.end()) {
-                        memoized_prefixes.emplace(prefix_t2, std::deque<value_type>());
-                    }
-                    std::deque<value_type>& buf = memoized_prefixes[prefix_t2];
-
-                    buf.insert(buf.end(), varr.begin(), varr.end());
-                }
-            }
-        }
-
-        for (auto const& item : memoized_prefixes) {
-            // copy the deque into a vector so we can sort without
-            // modifying the original array
-            intarray varr(item.second.begin(), item.second.end());
-
-            // delta-encode values, sorted in descending order.
-            std::sort(varr.begin(), varr.end(), std::greater<uint64_t>());
-
-            if (varr.size() > PREFIX_MAX_GRID_LENGTH) {
-                // for the prefix memos we're only going to ever use 500k max anyway
-                varr.resize(PREFIX_MAX_GRID_LENGTH);
-            }
-
-            packVec(varr, db, item.first);
-        }
-    } catch (std::exception const& ex) {
-        return Nan::ThrowTypeError(ex.what());
-    }
-    info.GetReturnValue().Set(Nan::Undefined());
-    return;
-}
 
 NAN_METHOD(RocksDBCache::pack) {
     if (info.Length() < 1) {
@@ -709,39 +458,6 @@ void mergeAfter(uv_work_t* req) {
     delete baton;
 }
 
-NAN_METHOD(MemoryCache::list) {
-    try {
-        Nan::Utf8String utf8_value(info[0]);
-        if (utf8_value.length() < 1) {
-            return Nan::ThrowTypeError("first arg must be a String");
-        }
-        MemoryCache* c = node::ObjectWrap::Unwrap<MemoryCache>(info.This());
-        Local<Array> ids = Nan::New<Array>();
-
-        unsigned idx = 0;
-        for (auto const& item : c->cache_) {
-            Local<Array> out = Nan::New<Array>();
-            out->Set(0, Nan::New(item.first.substr(0, item.first.find(LANGFIELD_SEPARATOR))).ToLocalChecked());
-
-            langfield_type langfield = extract_langfield(item.first);
-            if (langfield == ALL_LANGUAGES) {
-                out->Set(1, Nan::Null());
-            } else {
-                out->Set(1, langfieldToLangarray(langfield));
-            }
-
-            ids->Set(idx++, out);
-        }
-
-        info.GetReturnValue().Set(ids);
-        return;
-    } catch (std::exception const& ex) {
-        return Nan::ThrowTypeError(ex.what());
-    }
-    info.GetReturnValue().Set(Nan::Undefined());
-    return;
-}
-
 NAN_METHOD(RocksDBCache::list) {
     try {
         RocksDBCache* c = node::ObjectWrap::Unwrap<RocksDBCache>(info.This());
@@ -802,144 +518,6 @@ NAN_METHOD(RocksDBCache::merge) {
     return;
 }
 
-NAN_METHOD(MemoryCache::_set) {
-    if (info.Length() < 2) {
-        return Nan::ThrowTypeError("expected at least two info: id, data, [languages], [append]");
-    }
-    if (!info[0]->IsString()) {
-        return Nan::ThrowTypeError("first arg must be a String");
-    }
-    if (!info[1]->IsArray()) {
-        return Nan::ThrowTypeError("second arg must be an Array");
-    }
-    Local<Array> data = Local<Array>::Cast(info[1]);
-    if (data->IsNull() || data->IsUndefined()) {
-        return Nan::ThrowTypeError("an array expected for second argument");
-    }
-    try {
-
-        Nan::Utf8String utf8_id(info[0]);
-        if (utf8_id.length() < 1) {
-            return Nan::ThrowTypeError("first arg must be a String");
-        }
-        std::string id(*utf8_id);
-
-        langfield_type langfield;
-        if (info.Length() > 2 && !(info[2]->IsNull() || info[2]->IsUndefined())) {
-            if (!info[2]->IsArray()) {
-                return Nan::ThrowTypeError("third arg, if supplied must be an Array");
-            }
-            langfield = langarrayToLangfield(Local<Array>::Cast(info[2]));
-        } else {
-            langfield = ALL_LANGUAGES;
-        }
-
-        bool append = info.Length() > 3 && info[3]->IsBoolean() && info[3]->BooleanValue();
-
-        MemoryCache* c = node::ObjectWrap::Unwrap<MemoryCache>(info.This());
-        arraycache& arrc = c->cache_;
-        key_type key_id = static_cast<key_type>(id);
-        add_langfield(key_id, langfield);
-
-        arraycache::iterator itr2 = arrc.find(key_id);
-        if (itr2 == arrc.end()) {
-            arrc.emplace(key_id, intarray());
-        }
-        intarray& vv = arrc[key_id];
-
-        unsigned array_size = data->Length();
-        if (append) {
-            vv.reserve(vv.size() + array_size);
-        } else {
-            if (itr2 != arrc.end()) vv.clear();
-            vv.reserve(array_size);
-        }
-
-        for (unsigned i = 0; i < array_size; ++i) {
-            vv.emplace_back(static_cast<uint64_t>(data->Get(i)->NumberValue()));
-        }
-    } catch (std::exception const& ex) {
-        return Nan::ThrowTypeError(ex.what());
-    }
-    info.GetReturnValue().Set(Nan::Undefined());
-    return;
-}
-
-template <typename T>
-inline NAN_METHOD(_genericget) {
-    if (info.Length() < 1) {
-        return Nan::ThrowTypeError("expected at least one info: id, [languages]");
-    }
-    if (!info[0]->IsString()) {
-        return Nan::ThrowTypeError("first arg must be a String");
-    }
-    try {
-        Nan::Utf8String utf8_id(info[0]);
-        if (utf8_id.length() < 1) {
-            return Nan::ThrowTypeError("first arg must be a String");
-        }
-        std::string id(*utf8_id);
-
-        langfield_type langfield;
-        if (info.Length() > 1 && !(info[1]->IsNull() || info[1]->IsUndefined())) {
-            if (!info[1]->IsArray()) {
-                return Nan::ThrowTypeError("second arg, if supplied must be an Array");
-            }
-            langfield = langarrayToLangfield(Local<Array>::Cast(info[1]));
-        } else {
-            langfield = ALL_LANGUAGES;
-        }
-
-        T* c = node::ObjectWrap::Unwrap<T>(info.This());
-        intarray vector = __get(c, id, langfield);
-        if (!vector.empty()) {
-            std::size_t size = vector.size();
-            Local<Array> array = Nan::New<Array>(static_cast<int>(size));
-            for (uint32_t i = 0; i < size; ++i) {
-                array->Set(i, Nan::New<Number>(vector[i]));
-            }
-            info.GetReturnValue().Set(array);
-            return;
-        } else {
-            info.GetReturnValue().Set(Nan::Undefined());
-            return;
-        }
-    } catch (std::exception const& ex) {
-        return Nan::ThrowTypeError(ex.what());
-    }
-}
-
-NAN_METHOD(MemoryCache::_get) {
-    return _genericget<MemoryCache>(info);
-}
-
-NAN_METHOD(RocksDBCache::_get) {
-    return _genericget<RocksDBCache>(info);
-}
-
-NAN_METHOD(MemoryCache::New) {
-    if (!info.IsConstructCall()) {
-        return Nan::ThrowTypeError("Cannot call constructor as function, you need to use 'new' keyword");
-    }
-    try {
-        if (info.Length() < 1) {
-            return Nan::ThrowTypeError("expected 'id' argument");
-        }
-        if (!info[0]->IsString()) {
-            return Nan::ThrowTypeError("first argument 'id' must be a String");
-        }
-        MemoryCache* im = new MemoryCache();
-
-        im->Wrap(info.This());
-        info.This()->Set(Nan::New("id").ToLocalChecked(), info[0]);
-        info.GetReturnValue().Set(info.This());
-        return;
-    } catch (std::exception const& ex) {
-        return Nan::ThrowTypeError(ex.what());
-    }
-    info.GetReturnValue().Set(Nan::Undefined());
-    return;
-}
 
 NAN_METHOD(RocksDBCache::New) {
     if (!info.IsConstructCall()) {
@@ -983,67 +561,7 @@ NAN_METHOD(RocksDBCache::New) {
     return;
 }
 
-template <typename T>
-inline NAN_METHOD(_genericgetmatching) {
-    if (info.Length() < 2) {
-        return Nan::ThrowTypeError("expected two or three info: id, match_prefixes, [languages]");
-    }
-    if (!info[0]->IsString()) {
-        return Nan::ThrowTypeError("first arg must be a String");
-    }
-    if (!info[1]->IsBoolean()) {
-        return Nan::ThrowTypeError("second arg must be a Bool");
-    }
-    try {
-        Nan::Utf8String utf8_id(info[0]);
-        if (utf8_id.length() < 1) {
-            return Nan::ThrowTypeError("first arg must be a String");
-        }
-        std::string id(*utf8_id);
 
-        bool match_prefixes = info[1]->BooleanValue();
-
-        langfield_type langfield;
-        if (info.Length() > 2 && !(info[2]->IsNull() || info[2]->IsUndefined())) {
-            if (!info[2]->IsArray()) {
-                return Nan::ThrowTypeError("third arg, if supplied must be an Array");
-            }
-            langfield = langarrayToLangfield(Local<Array>::Cast(info[2]));
-        } else {
-            langfield = ALL_LANGUAGES;
-        }
-
-        T* c = node::ObjectWrap::Unwrap<T>(info.This());
-        intarray vector = __getmatching(c, id, match_prefixes, langfield);
-        if (!vector.empty()) {
-            std::size_t size = vector.size();
-            Local<Array> array = Nan::New<Array>(static_cast<int>(size));
-            for (uint32_t i = 0; i < size; ++i) {
-                auto obj = coverToObject(numToCover(vector[i]));
-
-                // these values don't make any sense outside the context of coalesce, so delete them
-                // it's a little clunky to set and then delete them, but this function as exposed
-                // to node is only used in debugging/testing, so, meh
-                obj->Delete(Nan::New("idx").ToLocalChecked());
-                obj->Delete(Nan::New("tmpid").ToLocalChecked());
-                obj->Delete(Nan::New("distance").ToLocalChecked());
-                obj->Delete(Nan::New("scoredist").ToLocalChecked());
-                array->Set(i, obj);
-            }
-            info.GetReturnValue().Set(array);
-            return;
-        } else {
-            info.GetReturnValue().Set(Nan::Undefined());
-            return;
-        }
-    } catch (std::exception const& ex) {
-        return Nan::ThrowTypeError(ex.what());
-    }
-}
-
-NAN_METHOD(MemoryCache::_getmatching) {
-    return _genericgetmatching<MemoryCache>(info);
-}
 
 NAN_METHOD(RocksDBCache::_getmatching) {
     return _genericgetmatching<RocksDBCache>(info);
