+++ src/binding.hpp
@@ -12,6 +12,11 @@
 #pragma clang diagnostic ignored "-Wold-style-cast"
 #pragma clang diagnostic ignored "-Wsign-conversion"
 #pragma clang diagnostic ignored "-Wshorten-64-to-32"
+
+#include "node_util.hpp"
+#include "coalesce.hpp"
+#include "memorycache.hpp"
+
 #include "radix_max_heap.h"
 #include "rocksdb/comparator.h"
 #include "rocksdb/db.h"
@@ -34,39 +39,8 @@
 
 namespace carmen {
 
-class noncopyable {
-  protected:
-    constexpr noncopyable() = default;
-    ~noncopyable() = default;
-    noncopyable(noncopyable const&) = delete;
-    noncopyable& operator=(noncopyable const&) = delete;
-};
+using namespace v8;
 
-typedef std::string key_type;
-typedef uint64_t value_type;
-typedef unsigned __int128 langfield_type;
-// fully cached item
-typedef std::vector<value_type> intarray;
-typedef std::vector<key_type> keyarray;
-typedef std::map<key_type, intarray> arraycache;
-
-class MemoryCache : public node::ObjectWrap {
-  public:
-    ~MemoryCache();
-    static Nan::Persistent<v8::FunctionTemplate> constructor;
-    static void Initialize(v8::Handle<v8::Object> target);
-    static NAN_METHOD(New);
-    static NAN_METHOD(pack);
-    static NAN_METHOD(list);
-    static NAN_METHOD(_get);
-    static NAN_METHOD(_getmatching);
-    static NAN_METHOD(_set);
-    static NAN_METHOD(coalesce);
-    explicit MemoryCache();
-    void _ref() { Ref(); }
-    void _unref() { Unref(); }
-    arraycache cache_;
-};
 
 class RocksDBCache : public node::ObjectWrap {
   public:
@@ -110,14 +84,8 @@ class NormalizationCache : public node::ObjectWrap {
 #define MEMO_PREFIX_LENGTH_T2 6
 #define PREFIX_MAX_GRID_LENGTH 500000
 
-constexpr langfield_type ALL_LANGUAGES = ~(langfield_type)(0);
-#define LANGFIELD_SEPARATOR '|'
-
-#define TYPE_MEMORY 1
-#define TYPE_ROCKSDB 2
-
-intarray __getmatching(MemoryCache const* c, std::string phrase, bool match_prefixes, langfield_type langfield);
 intarray __getmatching(RocksDBCache const* c, std::string phrase, bool match_prefixes, langfield_type langfield);
+intarray __get(RocksDBCache const* c, std::string phrase, langfield_type langfield);
 
 } // namespace carmen
