+++ src/node_util.cpp
