+++ src/node_util.hpp
@@ -2,7 +2,10 @@
 #define __CARMEN_NODE_UTIL_HPP__
 
 #include "binding.hpp"
+#include "memorycache.hpp"
 #include "cpp_util.hpp"
+
+#include <nan.h>
 
 namespace carmen {
 
@@ -36,6 +39,108 @@ inline Local<v8::Array> langfieldToLangarray(langfield_type langfield) {
         }
     }
     return langs;
+}
+
+template <typename T>
+inline NAN_METHOD(_genericget) {
+    if (info.Length() < 1) {
+        return Nan::ThrowTypeError("expected at least one info: id, [languages]");
+    }
+    if (!info[0]->IsString()) {
+        return Nan::ThrowTypeError("first arg must be a String");
+    }
+    try {
+        Nan::Utf8String utf8_id(info[0]);
+        if (utf8_id.length() < 1) {
+            return Nan::ThrowTypeError("first arg must be a String");
+        }
+        std::string id(*utf8_id);
+
+        langfield_type langfield;
+        if (info.Length() > 1 && !(info[1]->IsNull() || info[1]->IsUndefined())) {
+            if (!info[1]->IsArray()) {
+                return Nan::ThrowTypeError("second arg, if supplied must be an Array");
+            }
+            langfield = langarrayToLangfield(Local<Array>::Cast(info[1]));
+        } else {
+            langfield = ALL_LANGUAGES;
+        }
+
+        T* c = node::ObjectWrap::Unwrap<T>(info.This());
+        intarray vector = __get(c, id, langfield);
+        if (!vector.empty()) {
+            std::size_t size = vector.size();
+            Local<Array> array = Nan::New<Array>(static_cast<int>(size));
+            for (uint32_t i = 0; i < size; ++i) {
+                array->Set(i, Nan::New<Number>(vector[i]));
+            }
+            info.GetReturnValue().Set(array);
+            return;
+        } else {
+            info.GetReturnValue().Set(Nan::Undefined());
+            return;
+        }
+    } catch (std::exception const& ex) {
+        return Nan::ThrowTypeError(ex.what());
+    }
+}
+
+template <typename T>
+inline NAN_METHOD(_genericgetmatching) {
+    if (info.Length() < 2) {
+        return Nan::ThrowTypeError("expected two or three info: id, match_prefixes, [languages]");
+    }
+    if (!info[0]->IsString()) {
+        return Nan::ThrowTypeError("first arg must be a String");
+    }
+    if (!info[1]->IsBoolean()) {
+        return Nan::ThrowTypeError("second arg must be a Bool");
+    }
+    try {
+        Nan::Utf8String utf8_id(info[0]);
+        if (utf8_id.length() < 1) {
+            return Nan::ThrowTypeError("first arg must be a String");
+        }
+        std::string id(*utf8_id);
+
+        bool match_prefixes = info[1]->BooleanValue();
+
+        langfield_type langfield;
+        if (info.Length() > 2 && !(info[2]->IsNull() || info[2]->IsUndefined())) {
+            if (!info[2]->IsArray()) {
+                return Nan::ThrowTypeError("third arg, if supplied must be an Array");
+            }
+            langfield = langarrayToLangfield(Local<Array>::Cast(info[2]));
+        } else {
+            langfield = ALL_LANGUAGES;
+        }
+
+        T* c = node::ObjectWrap::Unwrap<T>(info.This());
+        intarray vector = __getmatching(c, id, match_prefixes, langfield);
+        if (!vector.empty()) {
+            std::size_t size = vector.size();
+            Local<Array> array = Nan::New<Array>(static_cast<int>(size));
+            for (uint32_t i = 0; i < size; ++i) {
+                auto obj = coverToObject(numToCover(vector[i]));
+
+                // these values don't make any sense outside the context of coalesce, so delete them
+                // it's a little clunky to set and then delete them, but this function as exposed
+                // to node is only used in debugging/testing, so, meh
+                obj->Delete(Nan::New("idx").ToLocalChecked());
+                obj->Delete(Nan::New("tmpid").ToLocalChecked());
+                obj->Delete(Nan::New("distance").ToLocalChecked());
+                obj->Delete(Nan::New("scoredist").ToLocalChecked());
+                array->Set(i, obj);
+            }
+            info.GetReturnValue().Set(array);
+            return;
+        } else {
+            info.GetReturnValue().Set(Nan::Undefined());
+            return;
+        }
+    } catch (std::exception const& ex) {
+        return Nan::ThrowTypeError(ex.what());
+    }
 }
 
 } // namespace carmen
